change: renamed plugin `grpc-proxy` to `grpc-transcode`.
