+++ conf/config.yaml
@@ -35,4 +35,4 @@ plugins:                          # plugin list
   - jwt-auth
   - zipkin
   - ip-restriction
-  - grpc-proxy
+  - grpc-transcode
