+++ doc/install-dependencies.md
@@ -94,8 +94,7 @@ CentOS 6
 
 TODO
 
-The compilation of `libr3` relies on later versions of `autoconf` and `pcre`, but the CentOS 6 comes with a lower version, will support CentOS 6 later.
-
+Failed to compile `rapidjson`, and the compilation of `libr3` relies on later versions of `autoconf` and `pcre`, but the CentOS 6 comes with a lower version, will support CentOS 6 later.
 
 Mac OSX
 =======
