+++ lua/apisix/plugins/grpc-transcode/response.lua
@@ -1,4 +1,4 @@
-local util   = require("apisix.plugins.grpc-proxy.util")
+local util   = require("apisix.plugins.grpc-transcode.util")
 local core   = require("apisix.core")
 local pb     = require("pb")
 local ngx    = ngx
