+++ t/debug-mode.t
@@ -46,7 +46,7 @@ loaded plugin and sort by priority: 1003 name: limit-conn
 loaded plugin and sort by priority: 1002 name: limit-count
 loaded plugin and sort by priority: 1001 name: limit-req
 loaded plugin and sort by priority: 1000 name: node-status
-loaded plugin and sort by priority: 506 name: grpc-proxy
+loaded plugin and sort by priority: 506 name: grpc-transcode
 loaded plugin and sort by priority: 500 name: prometheus
 loaded plugin and sort by priority: 0 name: example-plugin
 loaded plugin and sort by priority: -1000 name: zipkin
